pyproject.toml
[project]
name = "hackathon-parrotai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "parrot-olympe>=7.7.5",
]
